Remove lodash.foreach, lodash.reduce
We weren't using lodash.foreach and lodash.reduce was easy to replace
with a handrolled version. I usually wouldn't do that there was a
request for it (issue #20) because the extremely deep paths it creates
causes problems on windows. Also it's one less network connection to
make when installing, so that's always good.